encapsulated db code and schema CRUD spike pt 2
